% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{fsgraph-package}
\alias{fsgraph}
\alias{fsgraph-package}
\title{Generate and analyze Feature-Sample Graph.}
\description{
Description and analysis of  Feature-Sample Graph.
}
\details{
The only function you're likely to need from \pkg{fsgraph} is
\code{\link{makeGraph}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
}

}
