% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraph.R
\name{getMetaColOrder}
\alias{getMetaColOrder}
\title{Get column order for metadata.}
\usage{
getMetaColOrder(df)
}
\arguments{
\item{df}{metadata \code{data.frame}}
}
\value{
vector of column names
}
\description{
Get column order for metadata.
}
\examples{
md<-data.frame(class=c(rep('A',4),rep('B',4)),
subclass=paste0(c(rep('A',4),
rep('B',4)),rep(c(1,2),4)),
sample=paste0('sample',1:8))
getMetaColOrder(md)
}
