% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraph.R
\name{makeGraph}
\alias{makeGraph}
\title{Make FS graph.}
\usage{
makeGraph(featureMatrix, metadata, colOrder, metaStrength = 1000)
}
\arguments{
\item{featureMatrix}{matrix of feature-sample edge strength.}

\item{metadata}{metadata \code{data.frame} for metanodes and metaedges.}

\item{colOrder}{see in \code{\link{makeMetaNodes}}}

\item{metaStrength}{see \code{strength} in \code{\link{makeMetaNodes}}}
}
\value{
igraph graph object
}
\description{
Make FS graph.
}
