% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraph.R
\name{makeMetaNodes}
\alias{makeMetaNodes}
\title{Function creates two dataframes for metanodes and edges between them.}
\usage{
makeMetaNodes(df, colOrder, strength = 1000)
}
\arguments{
\item{df}{metadata data frame, describing relationship between metanodes.}

\item{colOrder}{vector of column names. Metanodes defined by unique names in
each column will be connected in order defined by \code{colOrder}.}

\item{strength}{strength of edge between two metadata. \code{strength} could
be ether one value, applied to each edge, or vector of length
\code{length(colOrder)-1} for each pair of metanodes types.}
}
\value{
list with two \code{data.frames} vertex and edges.
}
\description{
Metanodes are sample and nodes describing them. Unique name in each column in
\code{df} defines metanode of type, specified by column name. Metanodes
should form a forrest. Connections between metanodes defined by order of
columns in \code{colOrder}. Metanodes which names are in the same row in
neighbour column in \code{colOrder} will be connected by edge of the strength
from \code{strength} argument.
The column wihch name is the first in \code{colOrder} will contains roots of
trees in the forrest. Second column will be connected to first, third to
second and so on.
}
\details{
If \code{colOrder} is not provided order of columns in \code{df} will be
calculated by ordering columns according number of distinct values in it in
increasing order by function \code{\link{getMetaColOrder}}.
}
\seealso{
\code{\link{getMetaColOrder}} for ordering columns.
}
