% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraph.R
\name{makeMetaVertex}
\alias{makeMetaVertex}
\title{Create metavertices from metadata column.}
\usage{
makeMetaVertex(column, df)
}
\arguments{
\item{column}{column to take}

\item{df}{data.frame to analyse}
}
\value{
data.frame of vertices
}
\description{
Create metavertices from metadata column.
}
\seealso{
\code{\link[igraph]{graph_from_data_frame}}
}
